A class for the targets and limits of a fishery projection.

The desired targets, limits and time steps used in fishery projections can be specified by creating an object of class fwdControl.

Constructor for fwdControl objects. Bare bones man pages. Better to look at the vignettes and tutorials.

fwdControl constructor for data.frame and array

fwdControl constructor for data.frame and numeric

fwdControl constructor for data.frame and missing

fwdControl constructor for list and missing

fwdControl constructor for a series of list

fwdControl constructor for mising and missing

fwdControl(target, iters, ...)

G(...)

# S4 method for data.frame,array
fwdControl(target, iters, ...)

# S4 method for data.frame,numeric
fwdControl(target, iters, ...)

# S4 method for data.frame,missing
fwdControl(target, iters, ...)

# S4 method for list,missing
fwdControl(target, iters, ...)

# S4 method for list,list
fwdControl(target, iters, ...)

# S4 method for missing,missing
fwdControl(target, iters, ...)

Arguments

target

The target. Can be a data.frame, a list or missing.

iters

target The iters. Can be an array, a numeric or missing.

...

Something

Details

[...]

Slots

target

The table of quantities and time steps used as target (data.frame).

iters

The values and limits for each target quantity and time step (array).

FCB

The matrix describing which FLCatch of which FLFishery catches which FLBiol. A matrix with 3 columns: F, C, and B.

Validity

VALIDITY

Neque porro quisquam est qui dolorem ipsum.

Accessors

All slots in the class have accessor and replacement methods defined that allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot. A numeric vector can also be used when replacing FLQuant slots, and the vector will be used to substitute the values in the slot, but not its other attributes.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity. If an unnamed FLQuant object is provided, this is used for sizing but not stored in any slot.

See also

data.frame

Examples

# CREATE targets on fishing mortality ('f') by year

target <- data.frame(year=2000:2010, value=rlnorm(11), quantity='f')

fwc <- fwdControl(target=target)

#> Warning: provided 2 variables to replace 1 variables

# INSPECT fwdControl object

show(fwc)

#> An object of class "fwdControl"
#>  (step) year quant min value max
#>       1 2000     f  NA 0.247  NA
#>       2 2001     f  NA 1.291  NA
#>       3 2002     f  NA 0.087  NA
#>       4 2003     f  NA 0.994  NA
#>       5 2004     f  NA 1.862  NA
#>       6 2005     f  NA 3.153  NA
#>       7 2006     f  NA 0.162  NA
#>       8 2007     f  NA 0.781  NA
#>       9 2008     f  NA 0.783  NA
#>      10 2009     f  NA 0.754  NA
#>      11 2010     f  NA 0.575  NA

# Construct from data.frame and array
fcn <- fwdControl(data.frame(year=2000:2005, quant='f', value=0.5))
# Construct a fwdControl with some targets having multiple Biols, specified using the G() function
fwdControl(list(year=2000:2001, value=200, quant="catch", biol=G(1,2)),
  list(year=2002:2003, value=100, quant="catch", biol=c(1,2)))

#> An object of class "fwdControl"
#>  (step) year quant biol min   value max
#>       1 2000 catch 1, 2  NA 200.000  NA
#>       2 2001 catch 1, 2  NA 200.000  NA
#>       3 2002 catch    1  NA 100.000  NA
#>       4 2003 catch    2  NA 100.000  NA