Change names of biols, catches and fisheries in the control object into integer positions

Change names of biols, catches and fisheries in the control object into integer positions

match_posns_names(trg, biol_names, fishery_catch_names)

Arguments

trg
The target slot of a fwdControl object

biol_names
A vector of names in the FLBiols objects

fishery_catch_names
A named list - elements of list are vector of the catch names of each fishery

Value

The updated target slot

Details

Before calling the C++ code it is necessary for the catch, fishery and biol columns (and their Rel) equivalents to be integers. The user can specify by name. This function changes the name to integer position and throws an error if the name does not match.