Generate lists of randomly sized and filled FLQuant objects

Generate a list of FLQuant objects filled with normally distributed random numbers with a mean of 0. FLQuant objects can be randomly sized, depening on arguments passed to random_FLQuant_generator(). Used for automatic testing, particularly of the FLQuant7_base<T> class in CPP.

random_FLQuant_list_generator(min_elements = 1, max_elements = 10, ...)

Arguments

min_elements
The minimum number of elements in the list. Default is 1.

max_elements
The maximum number of elements in the list. Default is 10.

...
Other arguments to pass to random_FLQuant_generator(), e.g. those that fix the size of the objects.

Value

A list of FLQuant objects

Examples

flq_list <- random_FLQuant_list_generator()
length(flq_list)
#> [1] 9
summary(flq_list)
#>       Length Class   Mode   
#>  [1,] 1200   FLQuant numeric
#>  [2,]  576   FLQuant numeric
#>  [3,]  648   FLQuant numeric
#>  [4,]  432   FLQuant numeric
#>  [5,] 1152   FLQuant numeric
#>  [6,]  900   FLQuant numeric
#>  [7,]  192   FLQuant numeric
#>  [8,] 1200   FLQuant numeric
#>  [9,] 2304   FLQuant numeric
lapply(flq_list, summary)
#> An object of class "FLQuant" with:
#> dim  :  5 8 5 1 3 2 
#> quant:  age 
#> units:  0.383 
#> 
#> Min    :  -375.092 
#> 1st Qu.:  -62.98161 
#> Mean   :  -0.398561 
#> Median :  0.009049171 
#> 3rd Qu.:  63.53501 
#> Max    :  293.0777 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 6 4 1 2 4 
#> quant:  age 
#> units:  0.129 
#> 
#> Min    :  -302.2625 
#> 1st Qu.:  -78.5567 
#> Mean   :  -9.115739 
#> Median :  -6.485316 
#> 3rd Qu.:  65.29446 
#> Max    :  303.1069 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  1 6 3 3 4 3 
#> quant:  age 
#> units:  0.907 
#> 
#> Min    :  -309.0213 
#> 1st Qu.:  -64.76098 
#> Mean   :  -2.036323 
#> Median :  -2.499626 
#> 3rd Qu.:  59.33068 
#> Max    :  324.8002 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  2 4 3 2 3 3 
#> quant:  age 
#> units:  0.103 
#> 
#> Min    :  -263.8728 
#> 1st Qu.:  -60.83066 
#> Mean   :  6.711573 
#> Median :  12.04261 
#> 3rd Qu.:  72.49246 
#> Max    :  389.825 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  4 6 3 4 2 2 
#> quant:  age 
#> units:  0.628 
#> 
#> Min    :  -318.2162 
#> 1st Qu.:  -66.08703 
#> Mean   :  3.466572 
#> Median :  4.368771 
#> 3rd Qu.:  72.71964 
#> Max    :  352.4508 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  5 6 5 3 1 2 
#> quant:  age 
#> units:  0.849 
#> 
#> Min    :  -345.0837 
#> 1st Qu.:  -71.67478 
#> Mean   :  -2.356242 
#> Median :  -3.138855 
#> 3rd Qu.:  66.28986 
#> Max    :  344.2634 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  1 6 2 2 4 2 
#> quant:  age 
#> units:  0.956 
#> 
#> Min    :  -330.0258 
#> 1st Qu.:  -81.62905 
#> Mean   :  -10.74557 
#> Median :  -12.17632 
#> 3rd Qu.:  62.5251 
#> Max    :  222.1985 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  2 5 5 4 2 3 
#> quant:  age 
#> units:  1.28 
#> 
#> Min    :  -382.9153 
#> 1st Qu.:  -69.47737 
#> Mean   :  -5.718647 
#> Median :  -6.955052 
#> 3rd Qu.:  62.58148 
#> Max    :  357.5345 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  4 8 2 4 3 3 
#> quant:  age 
#> units:  0.738 
#> 
#> Min    :  -318.4072 
#> 1st Qu.:  -67.10049 
#> Mean   :  1.241501 
#> Median :  -2.93447 
#> 3rd Qu.:  68.70609 
#> Max    :  379.8256 
#> NAs    :  0 %
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 
#> [[3]]
#> NULL
#> 
#> [[4]]
#> NULL
#> 
#> [[5]]
#> NULL
#> 
#> [[6]]
#> NULL
#> 
#> [[7]]
#> NULL
#> 
#> [[8]]
#> NULL
#> 
#> [[9]]
#> NULL
#>