Prepare object for future projection

Similar to the old STF method in FLAssess. Extends the object by a number of years and fill in the life history characteristics by taking a mean of the last few years.

stf(object, ...)

# S4 method for FLStock
stf(
  object,
  nyears = 3,
  wts.nyears = 3,
  fbar.nyears = wts.nyears,
  f.rescale = FALSE,
  arith.mean = TRUE,
  na.rm = TRUE,
  end = dims(object)$maxyear + nyears,
  disc.nyears = wts.nyears
)

# S4 method for FLBiol
stf(
  object,
  nyears = 3,
  wts.nyears = 3,
  arith.mean = TRUE,
  na.rm = TRUE,
  end = dims(object)$maxyear + nyears
)

# S4 method for FLStocks
stf(object, ...)

Arguments

object
The object (FLStock or FLBiol)

...
Other things.

nyears
Number of years to extend the object

wts.nyears
Number of years to average over to get the future mean weights at age.

fbar.nyears
Number of years to average the F over (only used if object is an FLStock)

f.rescale
Rescale F (TRUE or FALSE - default is FALSE)

arith.mean
If TRUE the arithmetic mean is used. If FALSE the geometric mean is used. Default is TRUE.

na.rm
For the mean function.

end
My beautiful friend

disc.nyears
Number of years to average over to get the future mean proportion of discards at age.

Examples

data(ple4)
proj <- stf(ple4, 3)