Generate randomly sized and filled FLQuant objects

Generate a randomly or fixed sized FLQuant filled with normally distributed random numbers with a mean of 0. Used for automatic testing.

random_FLQuant_generator(
  fixed_dims = rep(NA, 6),
  min_dims = rep(1, 6),
  max_dims = pmax(min_dims, c(5, 10, 5, 4, 4, 5)),
  min_age_name = 1,
  sd = 100
)

Arguments

fixed_dims
A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).

min_dims
A vector of length 6 with minimum size of each of the FLQuant dimensions. Default value is c(1,1,1,1,1,1).

max_dims
A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,10,5,4,4,5).

min_age_name
The name of the first age group.

sd
The standard deviation of the random numbers. Passed to rnorm() Default is 100.

Value

An FLQuant

Examples

flq <- random_FLQuant_generator()
dim(flq)

#> [1] 3 5 3 3 2 1

summary(flq)

#> An object of class "FLQuant" with:
#> dim  :  3 5 3 3 2 1 
#> quant:  age 
#> units:  1.18 
#> 
#> Min    :  -256.173 
#> 1st Qu.:  -61.47061 
#> Mean   :  -2.781357 
#> Median :  -5.374132 
#> 3rd Qu.:  59.66756 
#> Max    :  379.7615 
#> NAs    :  0 %

flq <- random_FLQuant_generator(fixed_dims = c(NA,10,1,4,1,NA))
dim(flq)

#> [1]  3 10  1  4  1  1

summary(flq)

#> An object of class "FLQuant" with:
#> dim  :  3 10 1 4 1 1 
#> quant:  age 
#> units:  2.04 
#> 
#> Min    :  -236.2909 
#> 1st Qu.:  -59.97786 
#> Mean   :  1.378532 
#> Median :  -12.8695 
#> 3rd Qu.:  65.30138 
#> Max    :  277.3016 
#> NAs    :  0 %