Generate randomly sized and filled FLQuant objects

Generate a randomly or fixed sized FLQuant filled with normally distributed random numbers with a mean of 0. Used for automatic testing.

random_FLQuant_generator(
  fixed_dims = rep(NA, 6),
  min_dims = rep(1, 6),
  max_dims = pmax(min_dims, c(5, 10, 5, 4, 4, 5)),
  min_age_name = 1,
  sd = 100
)

Arguments

fixed_dims
A vector of length 6 with the fixed length of each of the FLQuant dimensions. If any value is NA it is randomly set using the max_dims argument. Default value is rep(NA,6).

min_dims
A vector of length 6 with minimum size of each of the FLQuant dimensions. Default value is c(1,1,1,1,1,1).

max_dims
A vector of length 6 with maximum size of each of the FLQuant dimensions. Default value is c(5,10,5,4,4,5).

min_age_name
The name of the first age group.

sd
The standard deviation of the random numbers. Passed to rnorm() Default is 100.

Value

An FLQuant

Examples

flq <- random_FLQuant_generator()
dim(flq)
#> [1] 3 6 3 3 3 2
summary(flq)
#> An object of class "FLQuant" with:
#> dim  :  3 6 3 3 3 2 
#> quant:  age 
#> units:  1.9 
#> 
#> Min    :  -325.9823 
#> 1st Qu.:  -70.25844 
#> Mean   :  0.6458963 
#> Median :  1.713319 
#> 3rd Qu.:  65.3906 
#> Max    :  360.9937 
#> NAs    :  0 %
flq <- random_FLQuant_generator(fixed_dims = c(NA,10,1,4,1,NA))
dim(flq)
#> [1]  1 10  1  4  1  4
summary(flq)
#> An object of class "FLQuant" with:
#> dim  :  1 10 1 4 1 4 
#> quant:  age 
#> units:  0.633 
#> 
#> Min    :  -255.4237 
#> 1st Qu.:  -65.96578 
#> Mean   :  -3.896104 
#> Median :  -4.49757 
#> 3rd Qu.:  53.15359 
#> Max    :  206.4432 
#> NAs    :  0 %