Generate lists of randomly sized and filled FLQuant objects

Generate a list of FLQuant objects filled with normally distributed random numbers with a mean of 0. FLQuant objects can be randomly sized, depening on arguments passed to random_FLQuant_generator(). Used for automatic testing, particularly of the FLQuant7_base<T> class in CPP.

random_FLQuant_list_generator(min_elements = 1, max_elements = 10, ...)

Arguments

min_elements
The minimum number of elements in the list. Default is 1.

max_elements
The maximum number of elements in the list. Default is 10.

...
Other arguments to pass to random_FLQuant_generator(), e.g. those that fix the size of the objects.

Value

A list of FLQuant objects

Examples

flq_list <- random_FLQuant_list_generator()
length(flq_list)

#> [1] 3

summary(flq_list)

#>      Length Class   Mode   
#> [1,] 360    FLQuant numeric
#> [2,] 126    FLQuant numeric
#> [3,]  48    FLQuant numeric

lapply(flq_list, summary)

#> An object of class "FLQuant" with:
#> dim  :  3 5 1 3 4 2 
#> quant:  age 
#> units:  0.125 
#> 
#> Min    :  -300.1746 
#> 1st Qu.:  -59.83575 
#> Mean   :  6.574699 
#> Median :  -0.06301166 
#> 3rd Qu.:  71.94314 
#> Max    :  313.7456 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 7 3 1 1 2 
#> quant:  age 
#> units:  0.00244 
#> 
#> Min    :  -356.1374 
#> 1st Qu.:  -70.29939 
#> Mean   :  -3.690934 
#> Median :  2.816598 
#> 3rd Qu.:  70.32917 
#> Max    :  198.6368 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 1 2 1 4 2 
#> quant:  age 
#> units:  0.196 
#> 
#> Min    :  -239.2456 
#> 1st Qu.:  -70.83193 
#> Mean   :  -12.49321 
#> Median :  -9.891053 
#> 3rd Qu.:  65.31259 
#> Max    :  151.845 
#> NAs    :  0 %

#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 
#> [[3]]
#> NULL
#>