Generate lists of randomly sized and filled FLQuant objects

Generate a list of FLQuant objects filled with normally distributed random numbers with a mean of 0. FLQuant objects can be randomly sized, depening on arguments passed to random_FLQuant_generator(). Used for automatic testing, particularly of the FLQuant7_base<T> class in CPP.

random_FLQuant_list_generator(min_elements = 1, max_elements = 10, ...)

Arguments

min_elements

The minimum number of elements in the list. Default is 1.

max_elements

The maximum number of elements in the list. Default is 10.

...

Other arguments to pass to random_FLQuant_generator(), e.g. those that fix the size of the objects.

Value

A list of FLQuant objects

Examples

flq_list <- random_FLQuant_list_generator()
length(flq_list)
#> [1] 9
summary(flq_list)
#>       Length Class   Mode   
#>  [1,] 1296   FLQuant numeric
#>  [2,]  432   FLQuant numeric
#>  [3,] 1920   FLQuant numeric
#>  [4,]  810   FLQuant numeric
#>  [5,]  768   FLQuant numeric
#>  [6,] 1620   FLQuant numeric
#>  [7,]  288   FLQuant numeric
#>  [8,] 1350   FLQuant numeric
#>  [9,]  768   FLQuant numeric
lapply(flq_list, summary)
#> An object of class "FLQuant" with:
#> dim  :  3 6 3 3 2 4 
#> quant:  age 
#> units:  0.293 
#> 
#> Min    :  -327.6326 
#> 1st Qu.:  -64.66422 
#> Mean   :  0.02826029 
#> Median :  -0.1698827 
#> 3rd Qu.:  65.81411 
#> Max    :  351.7334 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  2 8 3 3 3 1 
#> quant:  age 
#> units:  0.261 
#> 
#> Min    :  -306.0353 
#> 1st Qu.:  -53.56096 
#> Mean   :  2.048558 
#> Median :  1.183758 
#> 3rd Qu.:  59.67682 
#> Max    :  258.6045 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  1 8 5 4 4 3 
#> quant:  age 
#> units:  1.9 
#> 
#> Min    :  -391.518 
#> 1st Qu.:  -75.15779 
#> Mean   :  -4.135184 
#> Median :  -4.505855 
#> 3rd Qu.:  64.10013 
#> Max    :  271.5969 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 10 1 3 3 3 
#> quant:  age 
#> units:  1.49 
#> 
#> Min    :  -264.4179 
#> 1st Qu.:  -67.79442 
#> Mean   :  1.620565 
#> Median :  -3.615611 
#> 3rd Qu.:  74.21189 
#> Max    :  332.7672 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  2 8 4 1 3 4 
#> quant:  age 
#> units:  0.814 
#> 
#> Min    :  -295.7087 
#> 1st Qu.:  -65.34747 
#> Mean   :  2.083515 
#> Median :  6.997486 
#> 3rd Qu.:  74.32511 
#> Max    :  347.3052 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  5 6 2 3 3 3 
#> quant:  age 
#> units:  0.169 
#> 
#> Min    :  -363.976 
#> 1st Qu.:  -63.92527 
#> Mean   :  3.967118 
#> Median :  4.582356 
#> 3rd Qu.:  73.59307 
#> Max    :  311.5998 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 2 4 2 2 3 
#> quant:  age 
#> units:  1.78 
#> 
#> Min    :  -299.7302 
#> 1st Qu.:  -62.15069 
#> Mean   :  4.376122 
#> Median :  12.72197 
#> 3rd Qu.:  71.861 
#> Max    :  344.0222 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  3 5 2 3 3 5 
#> quant:  age 
#> units:  0.38 
#> 
#> Min    :  -284.5852 
#> 1st Qu.:  -68.56712 
#> Mean   :  -4.787184 
#> Median :  -5.129524 
#> 3rd Qu.:  64.01942 
#> Max    :  346.0432 
#> NAs    :  0 %
#> An object of class "FLQuant" with:
#> dim  :  2 2 4 4 3 4 
#> quant:  age 
#> units:  0.755 
#> 
#> Min    :  -366.0362 
#> 1st Qu.:  -67.34602 
#> Mean   :  -1.656538 
#> Median :  -3.287767 
#> 3rd Qu.:  65.01579 
#> Max    :  287.2658 
#> NAs    :  0 %
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#> 
#> [[3]]
#> NULL
#> 
#> [[4]]
#> NULL
#> 
#> [[5]]
#> NULL
#> 
#> [[6]]
#> NULL
#> 
#> [[7]]
#> NULL
#> 
#> [[8]]
#> NULL
#> 
#> [[9]]
#> NULL
#>